requirements.txt
amqp==5.1.1
autopep8==2.0.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==2.2.28
django-allauth==0.54.0
django-cors-headers==3.11.0
django-ftpserver==0.8.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-silk==4.4.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gprof2dot==2022.7.29
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
pycodestyle==2.11.1
pycparser==2.21
pyftpdlib==1.5.8
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
